chore: rename baseui popover to just popover
